requirements.txt
uvicorn
fastapi
pyyaml
sqlalchemy
psycopg2
jinja2
python-multipart
pytz